backend/requirements.txt
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.2
daphne==4.0.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==5.0.2
django-allauth==0.52.0
django-cors-headers==4.3.1
django-environ==0.12.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.2.0
fastjsonschema==2.21.1
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
oauthlib==3.2.2
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.8
prometheus_client==0.22.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python3-openid==3.2.0
pytz==2025.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.0
Send2Trash==1.8.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
zope.interface==7.2
